+++ src/Sentry/Extensibility/DisabledHub.cs
@@ -21,6 +21,7 @@ public class DisabledHub : IHub, IDisposable
 
     private DisabledHub()
     {
+        Logger = SentryStructuredLogger.CreateDisabled(this);
     }
 
     /// <summary>
@@ -228,4 +229,11 @@ public class DisabledHub : IHub, IDisposable
     /// No-Op.
     /// </summary>
     public SentryId LastEventId => SentryId.Empty;
+
+    /// <summary>
+    /// Disabled Logger.
+    /// <para>This API is experimental and it may change in the future.</para>
+    /// </summary>
+    [Experimental(Infrastructure.DiagnosticId.ExperimentalFeature)]
+    public SentryStructuredLogger Logger { get; }
 }