+++ src/Sentry/Internal/Hub.cs
@@ -66,6 +66,8 @@ internal class Hub : IHub, IDisposable
             PushScope();
         }
 
+        Logger = new SentryStructuredLogger(this, ScopeManager, options, _clock);
+
 #if MEMORY_DUMP_SUPPORTED
         if (options.HeapDumpOptions is not null)
         {
@@ -745,4 +747,7 @@ internal class Hub : IHub, IDisposable
     }
 
     public SentryId LastEventId => CurrentScope.LastEventId;
+
+    [Experimental(DiagnosticId.ExperimentalFeature)]
+    public SentryStructuredLogger Logger { get; }
 }